chore: Build whole workspace on macos
- Add some macos stubs to gui-smoke-test.
- Hide `ebpf-turn-router` binary functionality behind #[cfg(target_arch = "bpf")]
