+++ .github/actions/setup-rust/action.yml
@@ -20,7 +20,7 @@ outputs:
     description: Compilable packages for the current OS
     value: ${{
       (runner.os == 'Linux' && '--workspace') ||
-      (runner.os == 'macOS' && '--workspace --exclude ebpf-turn-router --exclude gui-smoke-test --exclude client-ffi') ||
+      (runner.os == 'macOS' && '--workspace') ||
       (runner.os == 'Windows' && '--workspace --exclude ebpf-turn-router --exclude apple-client-ffi --exclude client-ffi') }}
   test-packages:
     description: Testable packages for the current OS
