+++ rust/relay/ebpf-turn-router/Cargo.toml
@@ -4,6 +4,11 @@ version = "0.1.0"
 edition = { workspace = true }
 license = { workspace = true }
 
+[package.metadata.cargo-udeps.ignore]
+normal = ["aya-ebpf", "aya-log-ebpf", "ebpf-shared", "network-types"]
+build = ["which"]
+development = ["hex-literal", "ip-packet"]
+
 [[bin]]
 name = "ebpf-turn-router-main" # This needs to be different from the package name otherwise the build-script fails to differentiate between the directory it is built in and the actual binary.
 path = "src/main.rs"
